requirements.txt
tensorflow<2.11
streamlit
numpy<2.0
pillow
opencv-python
matplotlib
scipy
